refactor : Mapping 정리, setter injection 적용
