+++ src/main/java/techcourse/myblog/web/ArticleController.java
@@ -7,51 +7,44 @@ import org.springframework.web.bind.annotation.*;
 import techcourse.myblog.domain.Article;
 import techcourse.myblog.domain.ArticleRepository;
 
-import static techcourse.myblog.web.MethodType.POST;
-import static techcourse.myblog.web.MethodType.PUT;
+import static techcourse.myblog.domain.MethodType.PUT;
 
 @Controller
+@RequestMapping("/articles")
 public class ArticleController {
-    @Autowired
+
     private ArticleRepository articleRepository;
 
-    @GetMapping("/")
-    public String goIndex(Model model) {
-        model.addAttribute("articles", articleRepository.findAll());
-        return "index";
+    @Autowired
+    public void setArticleRepository(ArticleRepository articleRepository) {
+        this.articleRepository = articleRepository;
     }
 
-    @GetMapping("/writing")
-    public String goArticleEdit(Model model) {
-        model.addAttribute("method", POST);
-        return "article-edit";
-    }
-
-    @PostMapping("/articles")
+    @PostMapping
     public String createArticle(Article article) {
         articleRepository.saveArticle(article);
         return "redirect:/articles/" + article.getId();
     }
 
-    @GetMapping("/articles/{id}")
+    @GetMapping("/{id}")
     public String goDetailArticle(@PathVariable int id, Model model) {
         model.addAttribute("article", articleRepository.getArticleById(id));
         return "article";
     }
 
-    @PutMapping("/articles/{id}")
+    @PutMapping("/{id}")
     public String updateArticle(@PathVariable int id, Article updatedArticle) {
         articleRepository.updateArticle(updatedArticle);
         return "redirect:/articles/" + id;
     }
 
-    @DeleteMapping("/articles/{id}")
+    @DeleteMapping("/{id}")
     public String deleteArticle(@PathVariable int id) {
         articleRepository.removeArticleById(id);
         return "redirect:/";
     }
 
-    @GetMapping("/articles/{id}/edit")
+    @GetMapping("/{id}/edit")
     public String editArticle(@PathVariable int id, Model model) {
         model.addAttribute("article", articleRepository.getArticleById(id));
         model.addAttribute("method", PUT);
