+++ src/main/java/techcourse/myblog/web/MainController.java
@@ -1,0 +1,31 @@
+package techcourse.myblog.web;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.GetMapping;
+import techcourse.myblog.domain.ArticleRepository;
+
+import static techcourse.myblog.domain.MethodType.POST;
+
+@Controller
+public class MainController {
+    private ArticleRepository articleRepository;
+
+    @Autowired
+    public void setArticleRepository(ArticleRepository articleRepository) {
+        this.articleRepository = articleRepository;
+    }
+
+    @GetMapping("/")
+    public String goIndex(Model model) {
+        model.addAttribute("articles", articleRepository.findAll());
+        return "index";
+    }
+
+    @GetMapping("/writing")
+    public String goArticleEdit(Model model) {
+        model.addAttribute("method", POST);
+        return "article-edit";
+    }
+}
